make scoped, move to options. empty initialisation handed by options configuration
