+++ src/OrchardCore.Modules/OrchardCore.Media/Processing/MediaTokenOptions.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace OrchardCore.Media.Processing
 {
-    public class MediaTokenSettings
+    public class MediaTokenOptions
     {
         public byte[] HashKey { get; set; }
     }
